Update log export tools and readme
- Update the utlity program to handle both
  Cloudwatch and Splunk exports.

- Update the S3 file upload shell script.

- Update the readme.md for how to use the tool.

BB2-943
