+++ insights/scripts/upload_folders_to_s3.sh
@@ -1,6 +1,7 @@
 #!/usr/bin/bash
 #
-# Script to copy folders generated by the convert_logs_from_splunk.py script to the S3 target folder for BFD-Insights
+# Script to compress and copy folders generated by the convert_logs_to_json_for_insights_s3.py
+# program to the S3 target folder in BFD-Insights AWS S3.
 #
 # This copies all "dt=*/" type folders from the current PWD directory to the target S3 folder.
 #
@@ -8,15 +9,23 @@
 #
 # Example:
 #
-# $ sh ../upload_folders_to_s3.sh s3://<S3 bucket>/databases/bb2/events-global-state-apps/
+# $ sh ../upload_folders_to_s3.sh s3://<S3 bucket>/databases/bb2/
 #
 # NOTE: You must source the BFD AWS credentials in to your shell before running this!
 #
-for dir in $(ls -d dt\=*)
+for dir in $(ls -d events-*)
 do
+  echo "==="
+  echo GZIP FILES IN:  ${dir}
+  echo
+  find ${dir} -type f -exec gzip {} \;
+  echo
+  echo "==="
+  echo
   echo "==="
   echo COPYING TO S3: ${dir}
   echo
   aws s3 cp "${dir}" "${1}${dir}" --recursive
+  echo
   echo "==="
 done
